old-uniporn/requirements.txt
Django==2.0.4
requests
tqdm
bs4
pyquery
pillow
opencv-python
tensorflow